Move tests.lua to tests/runtime and update require
Rename tests.lua to tests/runtime.lua and change the require in control.lua from "tests" to "tests/runtime". Organizes test code under a tests directory and ensures the module is loaded from the new location.
